+++ rpc/src/main/java/com/walmartlabs/concord/rpc/SecretStoreService.java
@@ -1,0 +1,6 @@
+package com.walmartlabs.concord.rpc;
+
+public interface SecretStoreService {
+
+    String decryptString(String s) throws ClientException;
+}