+++ server/api/src/main/java/com/walmartlabs/concord/server/api/security/secret/SecretType.java
@@ -1,6 +1,19 @@
 package com.walmartlabs.concord.server.api.security.secret;
 
 public enum SecretType {
+
+    /**
+     * A SSH key pair.
+     */
     KEY_PAIR,
-    USERNAME_PASSWORD
+
+    /**
+     * An username and a password.
+     */
+    USERNAME_PASSWORD,
+
+    /**
+     * Binary data.
+     */
+    DATA
 }