+++ server/impl/src/main/java/com/walmartlabs/concord/server/security/secret/BinaryDataSecret.java
@@ -1,0 +1,14 @@
+package com.walmartlabs.concord.server.security.secret;
+
+public class BinaryDataSecret implements Secret {
+
+    private final byte[] data;
+
+    public BinaryDataSecret(byte[] data) {
+        this.data = data;
+    }
+
+    public byte[] getData() {
+        return data;
+    }
+}