+++ server/impl/src/main/java/com/walmartlabs/concord/server/security/secret/SecretUtils.java
@@ -1,7 +1,5 @@
 package com.walmartlabs.concord.server.security.secret;
 
-import com.walmartlabs.concord.server.api.security.secret.SecretType;
-
 import javax.crypto.Cipher;
 import javax.crypto.spec.SecretKeySpec;
 import java.security.GeneralSecurityException;
@@ -20,21 +18,6 @@ public final class SecretUtils {
         }
     }
 
-    public static Secret decrypt(SecretType type, byte[] input, byte[] password, byte[] salt) {
-        Function<byte[], ? extends Secret> deserializer;
-        switch (type) {
-            case KEY_PAIR:
-                deserializer = KeyPair::deserialize;
-                break;
-            case USERNAME_PASSWORD:
-                deserializer = UsernamePassword::deserialize;
-                break;
-            default:
-                throw new IllegalArgumentException("Unknown secret type: " + type);
-        }
-        return decrypt(deserializer, input, password, salt);
-    }
-
     public static <T extends Secret> byte[] encrypt(Function<T, byte[]> serializer, T k, byte[] password, byte[] hash) throws SecurityException {
         try {
             return encrypt(serializer.apply(k), password, hash);
@@ -43,17 +26,17 @@ public final class SecretUtils {
         }
     }
 
-    private static byte[] encrypt(byte[] input, byte[] password, byte[] salt) throws GeneralSecurityException {
+    public static byte[] encrypt(byte[] input, byte[] password, byte[] salt) throws GeneralSecurityException {
         Cipher c = init(password, salt, Cipher.ENCRYPT_MODE);
         return c.doFinal(input);
     }
 
-    private static byte[] decrypt(byte[] input, byte[] password, byte[] salt) throws GeneralSecurityException {
+    public static byte[] decrypt(byte[] input, byte[] password, byte[] salt) throws GeneralSecurityException {
         Cipher c = init(password, salt, Cipher.DECRYPT_MODE);
         return c.doFinal(input);
     }
 
-    private static byte[] hash(byte[] in, byte[] salt) throws NoSuchAlgorithmException {
+    public static byte[] hash(byte[] in, byte[] salt) throws NoSuchAlgorithmException {
         MessageDigest digest = MessageDigest.getInstance("MD5");
         digest.update(salt);
         return digest.digest(in);